+++ src/cli/run.rs
@@ -7,7 +7,7 @@ use tokio::{
     io::{stdin, AsyncReadExt as _},
 };
 
-use lune::Lune;
+use lune::Runtime;
 
 use super::utils::files::{discover_script_path_including_lune_dirs, strip_shebang};
 
@@ -41,7 +41,7 @@ impl RunCommand {
         };
 
         // Create a new lune object with all globals & run the script
-        let result = Lune::new()
+        let result = Runtime::new()
             .with_args(self.script_args)
             .run(&script_display_name, strip_shebang(script_contents))
             .await;