+++ src/lune/util/traits.rs
@@ -1,7 +1,7 @@
 use mlua::prelude::*;
 
 use super::formatting::format_label;
-use crate::LuneError;
+use crate::RuntimeError;
 
 pub trait LuaEmitErrorExt {
     fn emit_error(&self, err: LuaError);
@@ -10,6 +10,6 @@ pub trait LuaEmitErrorExt {
 impl LuaEmitErrorExt for Lua {
     fn emit_error(&self, err: LuaError) {
         // NOTE: LuneError will pretty-format this error
-        eprintln!("{}\n{}", format_label("error"), LuneError::from(err));
+        eprintln!("{}\n{}", format_label("error"), RuntimeError::from(err));
     }
 }